requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
apipkg==1.5
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
delayed-assert==0.3.2
dill==0.3.2
execnet==1.7.1
idna==2.10
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pendulum==2.1.2
pluggy==0.13.1
py==1.9.0
pydash==4.8.0
pyparsing==2.4.7
pytest==6.0.1
pytest-forked==1.3.0
pytest-reportportal==5.0.4
pytest-rerunfailures==9.0
pytest-xdist==2.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytzdata==2020.1
reportportal-client==5.0.5
requests==2.24.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
verify==1.1.1
webdriver-manager==3.2.1
